Revert "Revert "updatecode""
This reverts commit bf3eb61676414d4765a3448cab986e31f2773492.

+++ ManagementUI-BE/src/main/java/com/fsoft/entity/Form.java
@@ -10,17 +10,17 @@ import javax.persistence.Table;
 @Entity
 @Table(name="form")
 public class Form {
-	private int id;
+	private Long id;
 	private String nameForm;
 	private String content;
 	
 	@Id
 	@Column(name="id")
 	@GeneratedValue(strategy= GenerationType.IDENTITY)
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	@Column(name="nameForm")
@@ -37,7 +37,7 @@ public class Form {
 	public void setContent(String content) {
 		this.content = content;
 	}
-	public Form(int id, String nameForm) {
+	public Form(Long id, String nameForm) {
 		super();
 		this.id = id;
 		this.nameForm = nameForm;
